[Product-JIRA] - fix JAHIA-4384: Root categories "undeletable"
git-svn-id: svn://localhost/jahia/branches/JAHIA-6-0-BRANCH@1248 69c369ed-ed09-0410-a4a1-e19a46c23626

+++ core/src/main/java/org/jahia/ajax/gwt/engines/categories/server/CategoryServiceImpl.java
@@ -484,6 +484,12 @@ public class CategoryServiceImpl extends AbstractJahiaGWTServiceImpl implements 
         for (Category curChildCategory : childCategories) {
             recursiveDeleteCategory(curChildCategory);
         }
+
+        //  do not remove 'root' category
+        List<Category>  parentCategories = currentCategory.getParentCategories();
+        if(parentCategories == null || parentCategories.size() == 0 ){
+            return;
+        }
         currentCategory.delete();
     }
